refactor: extract _MatrixBackedRecommender base for shared recommend() logic
Pulls the duplicated recommend() body from SVD and the kNN base into a single
_MatrixBackedRecommender in base.py. Subclasses now supply only fit() (which
sets self._matrix) and _user_scores(user_id), the item-aligned score vector.

The base owns the unknown-user fast path, the seen-item -inf mask, and the
top-n / np.isfinite filter, so every matrix-backed algorithm follows one
contract — and any new algorithm with the same data shape (e.g. ALS, BPR)
will only need ~10 lines of glue.

No behavior change: all 59 existing tests pass unmodified.

Closes #35
