+++ src/recommender_systems/base.py
@@ -4,10 +4,9 @@ from __future__ import annotations
 
 from abc import ABC, abstractmethod
 from collections.abc import Hashable
-from typing import TYPE_CHECKING
 
-if TYPE_CHECKING:
-    import pandas as pd
+import numpy as np
+import pandas as pd
 
 __all__ = ["Recommender"]
 
@@ -51,3 +50,28 @@ class Recommender(ABC):
         list of Hashable
             Item ids ordered from most to least recommended.
         """
+
+
+class _MatrixBackedRecommender(Recommender):
+    """Recommend by ranking the user's row of a materialized user-item matrix.
+
+    Subclasses fill in :meth:`fit` (which must set ``self._matrix``) and
+    :meth:`_user_scores`, an item-aligned numpy vector. This base owns the
+    unknown-user fast path, the seen-item mask, and the top-``n`` filter, so
+    every matrix-backed algorithm shares one recommend contract.
+    """
+
+    def __init__(self) -> None:
+        self._matrix = pd.DataFrame()
+
+    @abstractmethod
+    def _user_scores(self, user_id: Hashable) -> np.ndarray:
+        """Return an item-aligned score vector for ``user_id``."""
+
+    def recommend(self, user_id: Hashable, n: int = 10) -> list[Hashable]:
+        if user_id not in self._matrix.index:
+            return []
+        scores = self._user_scores(user_id).copy()
+        scores[self._matrix.loc[user_id].to_numpy() > 0] = -np.inf
+        order = np.argsort(-scores)
+        return [self._matrix.columns[i] for i in order if np.isfinite(scores[i])][:n]
