+++ src/recommender_systems/neighborhood.py
@@ -9,7 +9,7 @@ import numpy as np
 import pandas as pd
 from sklearn.metrics.pairwise import cosine_similarity
 
-from recommender_systems.base import Recommender
+from recommender_systems.base import _MatrixBackedRecommender
 from recommender_systems.data import build_user_item_matrix
 
 __all__ = ["ItemKNN", "UserKNN"]
@@ -25,12 +25,12 @@ def _keep_top_k(similarity: np.ndarray, k: int) -> np.ndarray:
     return similarity
 
 
-class _NeighborhoodCF(Recommender):
-    """Shared fit/recommend for neighborhood collaborative filtering."""
+class _NeighborhoodCF(_MatrixBackedRecommender):
+    """Shared fit for neighborhood collaborative filtering."""
 
     def __init__(self, k: int = 20) -> None:
+        super().__init__()
         self.k = k
-        self._matrix = pd.DataFrame()
         self._similarity = np.empty((0, 0))
 
     def fit(self, ratings: pd.DataFrame) -> _NeighborhoodCF:
@@ -38,19 +38,8 @@ class _NeighborhoodCF(Recommender):
         self._similarity = _keep_top_k(self._similarity_of(self._matrix.to_numpy()), self.k)
         return self
 
-    def recommend(self, user_id: Hashable, n: int = 10) -> list[Hashable]:
-        if user_id not in self._matrix.index:
-            return []
-        scores = self._scores(user_id)
-        scores[self._matrix.loc[user_id].to_numpy() > 0] = -np.inf
-        order = np.argsort(-scores)
-        return [self._matrix.columns[i] for i in order if np.isfinite(scores[i])][:n]
-
     @abstractmethod
     def _similarity_of(self, matrix: np.ndarray) -> np.ndarray: ...
-
-    @abstractmethod
-    def _scores(self, user_id: Hashable) -> np.ndarray: ...
 
 
 class ItemKNN(_NeighborhoodCF):
@@ -59,7 +48,7 @@ class ItemKNN(_NeighborhoodCF):
     def _similarity_of(self, matrix: np.ndarray) -> np.ndarray:
         return np.asarray(cosine_similarity(matrix.T))
 
-    def _scores(self, user_id: Hashable) -> np.ndarray:
+    def _user_scores(self, user_id: Hashable) -> np.ndarray:
         return np.asarray(self._similarity @ self._matrix.loc[user_id].to_numpy())
 
 
@@ -69,6 +58,6 @@ class UserKNN(_NeighborhoodCF):
     def _similarity_of(self, matrix: np.ndarray) -> np.ndarray:
         return np.asarray(cosine_similarity(matrix))
 
-    def _scores(self, user_id: Hashable) -> np.ndarray:
+    def _user_scores(self, user_id: Hashable) -> np.ndarray:
         weights = self._similarity[self._matrix.index.get_loc(user_id)]
         return np.asarray(weights @ self._matrix.to_numpy() / (weights.sum() or 1.0))
