+++ src/recommender_systems/svd.py
@@ -8,13 +8,13 @@ import numpy as np
 import pandas as pd
 from sklearn.decomposition import TruncatedSVD
 
-from recommender_systems.base import Recommender
+from recommender_systems.base import _MatrixBackedRecommender
 from recommender_systems.data import build_user_item_matrix
 
 __all__ = ["SVD"]
 
 
-class SVD(Recommender):
+class SVD(_MatrixBackedRecommender):
     """Recommend from a low-rank reconstruction of the user-item matrix.
 
     Parameters
@@ -26,9 +26,9 @@ class SVD(Recommender):
     """
 
     def __init__(self, n_factors: int = 20, random_state: int | None = None) -> None:
+        super().__init__()
         self.n_factors = n_factors
         self.random_state = random_state
-        self._matrix = pd.DataFrame()
         self._predicted = np.empty((0, 0))
 
     def fit(self, ratings: pd.DataFrame) -> SVD:
@@ -39,10 +39,5 @@ class SVD(Recommender):
         self._predicted = factors @ model.components_
         return self
 
-    def recommend(self, user_id: Hashable, n: int = 10) -> list[Hashable]:
-        if user_id not in self._matrix.index:
-            return []
-        scores = self._predicted[self._matrix.index.get_loc(user_id)].copy()
-        scores[self._matrix.loc[user_id].to_numpy() > 0] = -np.inf
-        order = np.argsort(-scores)
-        return [self._matrix.columns[i] for i in order if np.isfinite(scores[i])][:n]
+    def _user_scores(self, user_id: Hashable) -> np.ndarray:
+        return np.asarray(self._predicted[self._matrix.index.get_loc(user_id)])
